backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
django-filter>=23.3
dj-database-url>=2.1
psycopg2-binary>=2.9
Pillow
whitenoise>=6.6
gunicorn>=21.2
drf-spectacular>=0.26
